add conversion.py
